Retry the entire screenshotStitcher call (#20770)
* Retry the entire screenshotStitcher call

* Go back to a single run

* Only retry for this specific error.  Post more information including the git issue link
