+++ x-pack/test/reporting/configs/chromium_api.js
@@ -21,7 +21,6 @@ export default async function ({ readConfigFile }) {
       serverArgs: [
         ...reportingApiConfig.kbnTestServer.serverArgs,
         `--xpack.reporting.capture.browser.type=chromium`,
-        `--logging.verbose=true`,
       ],
     },
   };
