+++ x-pack/test/reporting/configs/chromium_functional.js
@@ -21,7 +21,6 @@ export default async function ({ readConfigFile }) {
       serverArgs: [
         ...functionalConfig.kbnTestServer.serverArgs,
         `--xpack.reporting.capture.browser.type=chromium`,
-        `--logging.verbose=true`,
       ],
     },
   };
